+++ src/main.rs
@@ -1,7 +1,7 @@
 use std::ffi::OsStr;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use std::process::Command;
-use std::sync::{Arc, Mutex};
+use std::sync::mpsc::channel;
 use std::time::{Duration, Instant};
 
 use ignore::WalkBuilder;
@@ -14,80 +14,66 @@ mod cli;
 fn main() {
     let matches = cli::build().get_matches();
 
-    let pattern = {
-        let mut pattern = Vec::new();
-        pattern.append(
+    let patterns = {
+        let mut patterns = Vec::new();
+        patterns.append(
             &mut matches
                 .values_of("directory")
                 .unwrap_or_default()
                 .map(Pattern::new_directory)
                 .collect::<Vec<_>>(),
         );
-        pattern.append(
+        patterns.append(
             &mut matches
                 .values_of("file")
                 .unwrap_or_default()
                 .map(Pattern::new_file)
                 .collect::<Vec<_>>(),
         );
-        pattern
+        patterns
     };
     let recursive = matches.is_present("recursive");
     let command = matches
         .values_of_os("command")
         .map(std::iter::Iterator::collect::<Vec<_>>);
 
-    let already_used: Arc<Mutex<Vec<PathBuf>>> = Arc::new(Mutex::new(Vec::new()));
-
-    let walk = {
-        let already_used = already_used.clone();
+    let rx = {
+        let (tx, rx) = channel();
         WalkBuilder::new(".")
-            .filter_entry(move |d| {
-                let path = d.path();
-                if !path.is_dir() {
-                    return false;
-                }
-                if recursive {
-                    return true;
-                }
-                if is_parent_already_checked(&already_used.lock().unwrap(), path) {
-                    return false;
-                }
-                true
-            })
-            .build()
-            .filter_map(|d| match d {
-                Ok(d) => {
-                    if let Some(err) = d.error() {
-                        eprintln!("Warning for path {}: {}", d.path().display(), err);
+            .filter_entry(|d| d.path().is_dir())
+            .build_parallel()
+            .run(|| {
+                let patterns = patterns.clone();
+                let tx = tx.clone();
+                Box::new(move |entry| {
+                    match entry {
+                        Ok(d) => {
+                            let path = d.path();
+                            if let Some(err) = d.error() {
+                                eprintln!("Warning for path {}: {}", d.path().display(), err);
+                            }
+                            if check_dir_is_project(&patterns, path) {
+                                tx.send(path.to_path_buf()).expect("failed to send");
+                                if !recursive {
+                                    return ignore::WalkState::Skip;
+                                }
+                            }
+                        }
+                        Err(err) => eprintln!("Couldn't enter directory {}", err),
                     }
-                    Some(d)
-                }
-                Err(err) => {
-                    eprintln!("Couldn't enter directory {}", err);
-                    None
-                }
-            })
-            .filter(|d| check_dir_is_project(&pattern, d.path()))
+                    ignore::WalkState::Continue
+                })
+            });
+        rx
     };
 
-    for dir in walk {
-        let path = dir.path();
-        {
-            let mut already_used = already_used.lock().unwrap();
-            if !recursive && is_parent_already_checked(&already_used, path) {
-                // Should never happen with synchronous walker but maybe a parallel one is introduced in the future
-                continue;
-            }
-            already_used.push(path.to_path_buf());
-        }
-
+    for path in rx {
         println!("{}", path.display());
         // TODO: maybe check path.exists()
 
         if let Some(command) = &command {
             let start = Instant::now();
-            let status = generate_command(command, path)
+            let status = generate_command(command, &path)
                 .status()
                 .expect("failed to execute process");
             let took = start.elapsed();
@@ -106,11 +92,6 @@ where
     let mut c = Command::new(program);
     c.args(command).current_dir(working_dir).env("PAGER", "cat");
     c
-}
-
-fn is_parent_already_checked(already_used: &[PathBuf], path: &Path) -> bool {
-    path.ancestors()
-        .any(|p| already_used.contains(&p.to_path_buf()))
 }
 
 fn format_duration(duration: Duration) -> String {