+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `--list` shows a list of all the matching directories instead of running a command.
 - Check deep globs like `src/*.ts`.
 
+### Changed
+
+- Speed things up with a parallel file walker.
+
 ## [0.1.0] - 2022-02-03
 
 Initial release